platsupportports: Add tx2 clock driver
This clock driver is the exact same as the one in libplatsupport. We
move it here so as to avoid dependency issues in libplatsupport. The
extra file is just a table to translate libplatsupport clock IDs to the
IDs that the BPMP recognises. Also link against the libtx2bpmp library.