libtx2bpmp: Share initialised BPMP structures
There are now multiple interfaces that require access to the BPMP and it
isn't possible to have multiple copies of BPMP state as the BPMP's
private state must be shared. Thus, we now keep a global BPMP structure
and share that instead of requiring each user to initialise a private
structure. This commit also updates the tx2 clock driver to reflect
those changes.