libtx2bpmp: Change BPMP public interface
The BPMP interface was a little unique in that it did not follow the
structure of similar interfaces like those in libplatsupport where there
are function pointers with space to place a pointer to private data.
This presents integration issues with the interface registration service
in libplatsupport. Hence, this commit adapts the BPMP public interface
so that it makes it easier to work with the interface registration
service.