libplatsupportports: Look for BPMP interface
This commit changes the TX2 reset and clock drivers to try to look for
a BPMP interface instance in the interface registration service before
initialising the interface itself. This is mostly for streamlining the
process of being able to share the BPMP interface between drivers which
require access to the BPMP interface.