git: Remove .gitignore
We discourage the use of per-repo .gitignore and encourage people to use
personal .gitignore to tell Git to ignore specific files. There also
isn't much point to have a per-repo .gitignore in the first place as we
do not build artifacts inside the repositories but rather inside
dedicated build directories.