libsdhcdrivers: Fix `mmc_nth_irq()` function
There was a bug in the `mmc_nth_irq()` i.e. the function would always
return IRQ number only if the given index is 0.

This was caused by calling sdhc_init() from sdio_init() with wrong
parameters i.e. passing pointer the array of irq numbers with an offset
of the initialized device index, and claiming that this array's size
is 1.

Currently entire irq array is exposed via `mmv_nth_irq()`.

Furthermore, there is a mismatch between indexing of SDHC peripherals
and the their enumeration.

To avoid confusion and misconfiguration, the i.MX6 SDHC1 enum element
is now set to 1.